… .vscode/settings.json
… dist/hashids.js
… dist/hashids.js.map
… esm/index.js.map
… package.json
… tests/custom-alphabet.test.ts
… tests/types.d.ts
… yarn.lock