- [x] switch to use arrays of chars internally, rather than iterating the strings every time
- [x] tuned performance by switching to prepared RegExps when validating strings
- [x] added benchmarking to compare versions

Here's how performance looks with these changes:

```
1.2.2	 	decoding	 67,839 ops/sec ±0.32% (196736 samples)
2.1.0	 	decoding	 21,974 ops/sec ±0.34% (65133 samples)
this(dist)	decoding	 95,617 ops/sec ±0.29% (274244 samples)
this		decoding	 99,728 ops/sec ±0.29% (285774 samples)

1.2.2	 	encoding	 139,561 ops/sec ±0.23% (393444 samples)
2.1.0	 	encoding	 81,586 ops/sec ±0.31% (235682 samples)
this(dist) 	encoding	 210,082 ops/sec ±0.26% (575353 samples)
this		encoding	 213,644 ops/sec ±0.25% (583851 samples)
```

Not only fixes the regression from v2.1.0 (#79), looks significantly better than v1. 🎉