@niieani 2020-02-03
:tada: This PR is included in version 2.2.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/hashids/v/2.2.0)
- [GitHub release](https://github.com/niieani/hashids.js/releases/tag/2.2.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: